etl-automation-pipeline/requirements.txt
pandas
pyyaml
scikit-learn
sqlalchemy
